clean: 入力画面のコードを整理
